cli/src/test/kotlin/org/coner/trailer/cli/command/event/EventListCommandTest.kt
package org.coner.trailer.cli.command.event

import assertk.assertThat
import assertk.assertions.isEqualTo
import com.github.ajalt.clikt.core.context
import io.mockk.every
import io.mockk.impl.annotations.MockK
import io.mockk.junit5.MockKExtension
import io.mockk.verifySequence
import org.coner.trailer.TestEvents
import org.coner.trailer.cli.clikt.StringBufferConsole
import org.coner.trailer.cli.view.EventTableView
import org.coner.trailer.io.service.EventService
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.extension.ExtendWith
import org.kodein.di.DI
import org.kodein.di.bind
import org.kodein.di.instance

@ExtendWith(MockKExtension::class)
class EventListCommandTest {

    lateinit var command: EventListCommand

    @MockK lateinit var service: EventService
    @MockK lateinit var view: EventTableView

    lateinit var testConsole: StringBufferConsole

    @BeforeEach
    fun before() {
        testConsole = StringBufferConsole()
        command = EventListCommand(
            di = DI {
                bind<EventService>() with instance(service)
                bind<EventTableView>() with instance(view)
            }
        ).apply {
            context {
                console = testConsole
            }
        }
    }

    @Test
    fun `It should list events`() {
        val events = listOf(
            TestEvents.Lscc2019.points1,
            TestEvents.Lscc2019.points2,
            TestEvents.Lscc2019.points3
        )
        every { service.list() } returns events
        val viewRendered = "view rendered events"
        every { view.render(events) } returns viewRendered

        command.parse(emptyArray())

        verifySequence {
            service.list()
            view.render(events)
        }
        assertThat(testConsole.output).isEqualTo(viewRendered)
    }
}